extension playlist bugfix + forced return of remote playlist (with ID 1)
Remote playlist could be rely on a wrong playlist ID. Added listener for
remove so we know that the user tried to delete the needed playlist.
If playlist is created for the first time it could happen that it won't be
returned because banshee is not refreshing it's services quick enough. So
we're operating on the (fresh created) playlist and make sure we include it.